preprocess/requirements.txt
pypdf==6.4.0
azure-storage-blob

